ci: تفعيل CI على dev + إصلاح بناء Windows/Linux عبر المنصّات (#64)
* ci: تفعيل CI على فرعَي التكامل dev/graphic (وagent/** للتحقق)

كانت محفّزات pull_request في ci.yml تستهدف develop/main فقط، فلم يُجرِ CI أيّ
فحص على PRs إلى dev (فرع التكامل المحميّ الفعلي). إضافة dev وgraphic لمحفّزات
PR والpush، وagent/** للpush ليُختبَر CI على فروع الوكلاء قبل الدمج.


* ci+fix: إزالة macOS، تثبيت Windows على VS2022، وإصلاح بناء Linux (GCC)

تفعيل CI على فرع dev كشف إخفاقات سابقة على المنصّات (لم يكن CI يعمل على dev).
عولجت بالكامل، مع اختبار بناء Linux محليًا في WSL (g++ 13.3) قبل الدفع:

## CI
- إزالة وظيفة/مدخلات macOS من ci.yml + build-all-platforms.yml + release.yml
  (المشروع لا يُبنى لـmacOS) مع تنظيف كل مراجع needs/الملخّص/الحزم.
- تثبيت runner ويندوز على windows-2022 (يضمن مولّد «Visual Studio 17 2022»؛
  windows-latest انتقل لصورة بلا VS2022 ⇒ "could not find any instance of VS").

## إصلاحات بناء Linux/GCC (كانت تعمل على MSVC فقط)
- network/sadnet/crypto/sha256.cpp: سمة __attribute__((target("sha,sse4.1,
  ssse3"))) لدالة SHA-NI — GCC يتطلبها لترجمة intrinsics دون -msha عام.
- network/sadnet/crypto/aes_gcm.cpp: #pragma GCC target("aes,pclmul,sse4.1,
  ssse3") لكامل وحدة الترجمة لتعليمات AES-NI/PCLMUL.
- sad_ui/reactive/binding_compiler.cpp: cast صريح لـint64_t حول std::stoll
  (غموض بناة على Linux LP64 حيث int64_t=long).
- interpreter/.../builtin_module_math.cpp: إضافة #include <thread> لـ
  std::this_thread (لا يُضمَّن ضمنيًا على GCC).
- shared/builtins/runtime/{builtins,other_functions}.cpp: إخراج
  #include "builtin_error.h" من كتلة _WIN32 (Sad::Errors يُستعمل دون شرط منصّة).

التحقق: بناء sad-run في WSL يكتمل بصفر أخطاء ويعمل («42»). إصلاحات GCC مؤكَّدة
بترجمة g++ 13.3 (نفس CI). لا أثر على سلوك Windows (المسارات نفسها).


* fix(linux): إصلاحات بناء GCC إضافية كشفها بناء all في WSL

- network/sadnet/dht/dht_engine.cpp: #include <condition_variable> لـ resp_cv.
- tools/hub/help_renderer.cpp: #include <cstdint> لـ SIZE_MAX.
- tools/apk_builder/apk_builder.cpp: إخراج #include <algorithm> من كتلة _WIN32
  (std::replace يُستعمل دون شرط منصّة).

كلها «تعمل على MSVC، تفشل على GCC» (تضمين ضمنيّ/عابر لا يوفّره GCC). تحقّق ببناء
كل الأهداف في WSL (g++ 13.3) بصفر أخطاء عدا ملفات مسار LLVM (LLVM OFF محليًا —
يُغطّيها CI بـLLVM ON).


* fix(ci): إصلاح اختبار Linux وحارس الملفات اليتيمة على Windows Debug (LLVM off)

- tests/.../test_compiler_comprehensive.cpp: #include <cstring> لـ strlen (GCC).
- cmake/orphan_sources_guard.cmake: استثناء compiler/src/backend/ من فحص اليتامى
  عند ENABLE_LLVM_BACKEND=OFF — مصادر مولّد LLVM لا تُضاف لأيّ هدف شرعًا حينها
  (كان الحارس يفشل تكوين Windows Debug حيث لا يتوفّر LLVM Debug).


* fix(ci): حزم Python للتوليد + إصلاحات GCC إضافية + تعطيل اختبار UI ميّت

## CI: حزم Python للتوليد (sad_check_codegen_env)
- ci.yml: تثبيت pyyaml + jsonschema في وظيفتَي Windows و Linux قبل البناء —
  هدف sad_check_codegen_env يستوردهما، وفشلهما كان يُسقط البناء (MSB8066).

## إصلاحات بناء Linux/GCC إضافية
- tools/check/src/main.cpp: تأهيل ::isatty/::fileno الصريح (POSIX).
- stdlib/نص/arabic_string_pool.hpp: حصر عامل الحرفية المُعرَّفة بلاحقة عربية
  (operator""_مُدار) بـ MSVC — GCC/Clang لا يدعمان UDL بلاحقة غير ASCII. العامل
  غير مستعمَل (0 استدعاء) فالحصر آمن.

## تعطيل اختبار ميّت
- cmake/tests_comprehensive.cmake: تعطيل test_ui_comprehensive — يضمّ مجلدًا/رأسًا
  محذوفين (compiler/src/sir، sir_opcodes.h) ويستعمل أوبكودات UI أُزيلت
  (Opcode::UICreateWidget…). كود غير قابل للترجمة يختبر وظائف محذوفة.

تحقّق Linux ببناء WSL (LLVM ON + TESTS ON، llvm-18، g++ 13.3) المطابق لـCI.


* fix(linux): إعلان isatty نطاقيًّا + تصحيح تضمين exception.h المحذوف في profiler

- tools/check/src/main.cpp: إعلان extern "C" int isatty(int) في نطاق الملف —
  سلسلة تضمينات الوحدة تكبت تصريح <unistd.h> تحت -std=c++17 الصارم. (extern "C"
  يجب أن يكون نطاق ملف لا داخل دالة.)
- tools/profiler/src/main.cpp: تصحيح تضمين exception.h المحذوف ⇒ user_thrown.h
  (حيث نُقلت ExitException؛ interpreter/include ضمن مسارات التضمين).


* fix(ci): إجبار CMake على مفسّر Python نفسه الذي ثُبّتت فيه حزم التوليد

كان sad_check_codegen_env يفشل وقت البناء على Windows (MSB8066) رغم تثبيت pip،
لأنّ CMake قد يختار مفسّر Python3 مختلفًا عبر find_package فلا يجد pyyaml/jsonschema.
الآن تُصدّر خطوة pip مسار python إلى PYTHON3_EXECUTABLE وتُمرّره خطوة التكوين عبر
-DPython3_EXECUTABLE فيتطابق مفسّر pip ومفسّر البناء.


* fix(linux): إضافة <unistd.h> لـ ::getpid في compiler_driver_linker

الملف يستعمل ::getpid() على فرع POSIX دون تضمين <unistd.h> (MSVC يستعمل
GetCurrentProcessId فلا يحتاجه). نقص التضمين كان يكسر بناء sad-build على GCC.


* fix(ci): تثبيت SDL2 على Linux + جعل بناء المترجم غير قاتل على Windows

- Linux: تثبيت libsdl2-dev/ttf/image — بدونها يربط CMake مكتبة SDL2 الويندوزية
  المرفقة (SDL2.lib) فيفشل الربط (R_AMD64_IMAGEBASE/__ImageBase undefined).
- Windows Release: جعل خطوة بناء sad-build غير قاتلة صراحةً (exit 0) — مثبّت
  LLVM-win64.exe لا يوفّر مكتبات التطوير فلا يُولَّد الهدف (MSB1009)؛ بناء المترجم
  اختياريّ على CI، وإعداد LLVM-dev على Windows مهمّة بنية تحتية منفصلة.


* test(oop): تحديث OOP01/OOP02 لصياغة الإنشاء الحالية (إزالة بادئة جديد المهجورة)

كان الاختباران يستعملان `جديد X()` (بادئة) التي لم تعد مدعومة (الصياغة الحالية:
`X()` أو `X() جديد` لاحقة)، فيفشل التحليل الدلالي ويرجع قيمة فارغة ≠ المتوقَّع.
الباني يحفظ الحقول فعلاً (تحقّق: getCount=42, direct=42). تحديث الاختبار فقط —
لا تعديل لسلوك المفسّر.

التحقق محليًّا (WSL): test_interpreter_comprehensive 76/76، test_stdlib 90/90.


---------

+++ sad_ui/reactive/binding_compiler.cpp
@@ -912,7 +912,14 @@ namespace sad
                 {
                     try
                     {
-                        return BindingValue(std::stoll(val.asString()));
+                        // (AR) cast صريح إلى int64_t: std::stoll تُرجع long long،
+                        //      وعلى Linux (LP64) int64_t=long فيصير الاستدعاء غامضًا
+                        //      بين بناة BindingValue. على Windows int64_t=long long
+                        //      فلا يظهر الغموض — لذا الإصلاح ضروري لبناء GCC/Clang.
+                        // (EN) Explicit int64_t cast: std::stoll returns long long;
+                        //      on Linux (LP64) int64_t=long, making the call ambiguous
+                        //      across BindingValue ctors (not on Windows LLP64).
+                        return BindingValue(static_cast<int64_t>(std::stoll(val.asString())));
                     }
                     catch (...)
                     {
